Refactor imports and remove deprecated typings module
- Updated import paths for typings to use the new structure in opticalib.core.
- Removed the typings.py module as it is no longer needed.
- Adjusted relevant files to reflect the new import paths for type aliases.
- Cleaned up factory functions by moving to the appropriate factory module.
