+++ opticalib/analyzer/__init__.py
@@ -1,3 +1,3 @@
-from .images_processing import *
+from .image_processing import *
 from .signals import *
 from .timeseries import *
