+++ opticalib/devices/wfs.py
@@ -1,13 +1,13 @@
-from . import _API as _api
+from ._API.base_devices import BaseWavefrontSensor
 
 
-class Ingot(_api.BaseWavefrontSensor): ...
+class Ingot(BaseWavefrontSensor): ...
 
 
-class ShackHartmann(_api.BaseWavefrontSensor): ...
+class ShackHartmann(BaseWavefrontSensor): ...
 
 
-class Pyramid(_api.BaseWavefrontSensor): ...
+class Pyramid(BaseWavefrontSensor): ...
 
 
-class BiOEdge(_api.BaseWavefrontSensor): ...
+class BiOEdge(BaseWavefrontSensor): ...
