+++ opticalib/ground/__init__.py
@@ -27,4 +27,4 @@ Contents
 - zernike.py : module for Zernike polynomials computation and fitting on images.
 """
 
-from . import roi, modal_decomposer, geometry, osutils, reconstructor
+from . import reconstructor, roi, modal_decomposer, geometry, osutils
